# feat(fmt): Improve & speedup code formatting