--- a/.github/workflows/sast-linters.yml
+++ b/.github/workflows/sast-linters.yml
@@ -69,11 +69,10 @@ jobs:
           go-version: '1.20'
 
       - name: Run golangci-lint
-        uses: golangci/golangci-lint-action@v3
+        uses: golangci/golangci-lint-action@v6
         with:
-          version: v1.54
+          version: v1.59
           skip-cache: true
-          args: --out-format=json
 
   nosec_alert:
     runs-on: ubuntu-22.04