--- a/.golangci.yml
+++ b/.golangci.yml
@@ -1,3 +1,8 @@
+run:
+  go: '1.20' # as in go.mod
+  timeout: 5m
+  tests: false # exclude test from linting
+
 linters:
   disable-all: true
   enable:
@@ -21,16 +26,17 @@ linters:
     - gci
     - gofmt
     - whitespace
+
 linters-settings:
   gocyclo:
     min-complexity: 11
   errcheck:
-    ignore: fmt:.*,io/ioutil:^Read.*,github.com/spf13/cobra:MarkFlagRequired,github.com/spf13/viper:BindPFlag
-  golint:
-    min-confidence: 1.1
+    exclude-functions:
+      - fmt:.*
+      - io/ioutil:^Read.*,
+      - github.com/spf13/cobra:MarkFlagRequired
+      - github.com/spf13/viper:BindPFlag
   exhaustive:
-    check-generated: false
-  exhaustivestruct:
     check-generated: false
   gci:
     # Gci controls Go package import order and makes it always deterministic
@@ -42,12 +48,7 @@ linters-settings:
     skip-generated: true
 
 issues:
+  exclude-generated: strict
   exclude:
     - composite
-run:
-  tests: false
-  deadline: 15m
-  timeout: 5m
-  skip-files:
-    - ".*\\.pb\\.go$"
-
+  exclude-dirs: [ ".git", ".github" ]