--- a/cmd/zetaclientd/debug.go
+++ b/cmd/zetaclientd/debug.go
@@ -187,7 +187,6 @@ func DebugCmd() *cobra.Command {
 				if err != nil {
 					return err
 				}
-
 			}
 			fmt.Println("BallotIdentifier : ", ballotIdentifier)