--- a/cmd/zetaclientd/main.go
+++ b/cmd/zetaclientd/main.go
@@ -42,5 +42,4 @@ func SetupConfigForTest() {
 	})
 
 	rand.Seed(time.Now().UnixNano())
-
 }