--- a/cmd/zetacored/main.go
+++ b/cmd/zetacored/main.go
@@ -18,7 +18,6 @@ func main() {
 	rootCmd, _ := NewRootCmd()
 
 	if err := svrcmd.Execute(rootCmd, "", app.DefaultNodeHome); err != nil {
-
 		switch e := err.(type) {
 		case server.ErrorCode:
 			os.Exit(e.Code)