--- a/e2e/runner/setup_zeta.go
+++ b/e2e/runner/setup_zeta.go
@@ -197,7 +197,6 @@ func (runner *E2ERunner) SetZEVMContracts() {
 	}
 	runner.ContextAppAddr = contextAppAddr
 	runner.ContextApp = contextApp
-
 }
 
 // SetupETHZRC20 sets up the ETH ZRC20 in the runner from the values queried from the chain