--- a/x/crosschain/client/cli/cli_tss.go
+++ b/x/crosschain/client/cli/cli_tss.go
@@ -19,7 +19,6 @@ func CmdUpdateTss() *cobra.Command {
 		Short: "Create a new TSSVoter",
 		Args:  cobra.ExactArgs(1),
 		RunE: func(cmd *cobra.Command, args []string) error {
-
 			argsPubkey, err := cast.ToStringE(args[0])
 			if err != nil {
 				return err
@@ -45,7 +44,6 @@ func CmdMigrateTssFunds() *cobra.Command {
 		Short: "Migrate TSS funds to the latest TSS address",
 		Args:  cobra.ExactArgs(2),
 		RunE: func(cmd *cobra.Command, args []string) error {
-
 			argsChainID, err := strconv.ParseInt(args[0], 10, 64)
 			if err != nil {
 				return err