--- a/x/crosschain/client/cli/cli_zeta_height.go
+++ b/x/crosschain/client/cli/cli_zeta_height.go
@@ -14,7 +14,6 @@ func CmdLastZetaHeight() *cobra.Command {
 		Short: "Query last Zeta Height",
 		Args:  cobra.ExactArgs(0),
 		RunE: func(cmd *cobra.Command, _ []string) error {
-
 			clientCtx, err := client.GetClientTxContext(cmd)
 			if err != nil {
 				return err