--- a/x/crosschain/client/querytests/suite.go
+++ b/x/crosschain/client/querytests/suite.go
@@ -52,7 +52,6 @@ func (s *CliTestSuite) SetupSuite() {
 	s.network = net
 	_, err = s.network.WaitForHeight(1)
 	s.Require().NoError(err)
-
 }
 
 func (s *CliTestSuite) TearDownSuite() {