--- a/x/crosschain/keeper/events.go
+++ b/x/crosschain/keeper/events.go
@@ -56,7 +56,6 @@ func EmitZetaWithdrawCreated(ctx sdk.Context, cctx types.CrossChainTx) {
 	if err != nil {
 		ctx.Logger().Error("Error emitting ZetaWithdrawCreated :", err)
 	}
-
 }
 
 func EmitOutboundSuccess(ctx sdk.Context, valueReceived string, oldStatus string, newStatus string, cctxIndex string) {
@@ -70,7 +69,6 @@ func EmitOutboundSuccess(ctx sdk.Context, valueReceived string, oldStatus string
 	if err != nil {
 		ctx.Logger().Error("Error emitting MsgVoteOutbound :", err)
 	}
-
 }
 
 func EmitOutboundFailure(ctx sdk.Context, valueReceived string, oldStatus string, newStatus string, cctxIndex string) {