--- a/x/crosschain/keeper/msg_server_update_tss.go
+++ b/x/crosschain/keeper/msg_server_update_tss.go
@@ -60,7 +60,6 @@ func (k msgServer) UpdateTssAddress(
 				migratorTx.CctxStatus.Status.String(),
 			)
 		}
-
 	}
 
 	k.GetObserverKeeper().SetTssAndUpdateNonce(ctx, tss)