--- a/x/emissions/client/cli/query_list_balances.go
+++ b/x/emissions/client/cli/query_list_balances.go
@@ -14,7 +14,6 @@ func CmdListPoolAddresses() *cobra.Command {
 		Short: "Query list-pool-addresses",
 		Args:  cobra.ExactArgs(0),
 		RunE: func(cmd *cobra.Command, _ []string) (err error) {
-
 			clientCtx, err := client.GetClientTxContext(cmd)
 			if err != nil {
 				return err