--- a/x/fungible/client/cli/tx_update_contract_bytecode.go
+++ b/x/fungible/client/cli/tx_update_contract_bytecode.go
@@ -15,7 +15,6 @@ func CmdUpdateContractBytecode() *cobra.Command {
 		Short: "Broadcast message UpdateContractBytecode",
 		Args:  cobra.ExactArgs(2),
 		RunE: func(cmd *cobra.Command, args []string) (err error) {
-
 			clientCtx, err := client.GetClientTxContext(cmd)
 			if err != nil {
 				return err