--- a/x/lightclient/client/cli/tx_enable_verification_flags.go
+++ b/x/lightclient/client/cli/tx_enable_verification_flags.go
@@ -24,7 +24,6 @@ func CmdEnableVerificationFlags() *cobra.Command {
 				`,
 		Args: cobra.ExactArgs(1),
 		RunE: func(cmd *cobra.Command, args []string) (err error) {
-
 			clientCtx, err := client.GetClientTxContext(cmd)
 			if err != nil {
 				return err