--- a/x/lightclient/keeper/grpc_query_header_enabled_chains.go
+++ b/x/lightclient/keeper/grpc_query_header_enabled_chains.go
@@ -51,5 +51,4 @@ func (k Keeper) HeaderEnabledChains(
 	}
 
 	return &types.QueryHeaderEnabledChainsResponse{HeaderEnabledChains: val.GetHeaderEnabledChains()}, nil
-
 }