--- a/x/observer/client/cli/query_blame.go
+++ b/x/observer/client/cli/query_blame.go
@@ -49,7 +49,6 @@ func CmdGetAllBlameRecords() *cobra.Command {
 		Short: "Query AllBlameRecords",
 		Args:  cobra.ExactArgs(0),
 		RunE: func(cmd *cobra.Command, _ []string) (err error) {
-
 			clientCtx, err := client.GetClientTxContext(cmd)
 			if err != nil {
 				return err