--- a/x/observer/client/cli/query_chain_params.go
+++ b/x/observer/client/cli/query_chain_params.go
@@ -48,7 +48,6 @@ func CmdGetChainParams() *cobra.Command {
 		Short: "Query GetChainParams",
 		Args:  cobra.ExactArgs(0),
 		RunE: func(cmd *cobra.Command, _ []string) (err error) {
-
 			clientCtx, err := client.GetClientTxContext(cmd)
 			if err != nil {
 				return err