--- a/x/observer/client/cli/query_get_tss_address.go
+++ b/x/observer/client/cli/query_get_tss_address.go
@@ -16,7 +16,6 @@ func CmdGetTssAddress() *cobra.Command {
 		Short: "Query current tss address",
 		Args:  cobra.MaximumNArgs(1),
 		RunE: func(cmd *cobra.Command, args []string) (err error) {
-
 			clientCtx, err := client.GetClientTxContext(cmd)
 			if err != nil {
 				return err
@@ -51,7 +50,6 @@ func CmdGetTssAddressByFinalizedZetaHeight() *cobra.Command {
 		Short: "Query tss address by finalized zeta height (for historical tss addresses)",
 		Args:  cobra.MaximumNArgs(2),
 		RunE: func(cmd *cobra.Command, args []string) (err error) {
-
 			clientCtx, err := client.GetClientTxContext(cmd)
 			if err != nil {
 				return err