--- a/x/observer/client/cli/query_observers.go
+++ b/x/observer/client/cli/query_observers.go
@@ -14,7 +14,6 @@ func CmdObserverSet() *cobra.Command {
 		Short: "Query observer set",
 		Args:  cobra.ExactArgs(0),
 		RunE: func(cmd *cobra.Command, _ []string) (err error) {
-
 			clientCtx, err := client.GetClientTxContext(cmd)
 			if err != nil {
 				return err