--- a/x/observer/client/cli/tx_add_observer.go
+++ b/x/observer/client/cli/tx_add_observer.go
@@ -18,7 +18,6 @@ func CmdAddObserver() *cobra.Command {
 		Short: "Broadcast message add-observer",
 		Args:  cobra.ExactArgs(3),
 		RunE: func(cmd *cobra.Command, args []string) (err error) {
-
 			clientCtx, err := client.GetClientTxContext(cmd)
 			if err != nil {
 				return err