--- a/x/observer/client/cli/tx_disable_cctx_flags.go
+++ b/x/observer/client/cli/tx_disable_cctx_flags.go
@@ -17,7 +17,6 @@ func CmdDisableCCTX() *cobra.Command {
 		Short: "Disable inbound and outbound for CCTX",
 		Args:  cobra.ExactArgs(2),
 		RunE: func(cmd *cobra.Command, args []string) (err error) {
-
 			clientCtx, err := client.GetClientTxContext(cmd)
 			if err != nil {
 				return err