--- a/x/observer/client/cli/tx_update_chain_params.go
+++ b/x/observer/client/cli/tx_update_chain_params.go
@@ -19,7 +19,6 @@ func CmdUpdateChainParams() *cobra.Command {
 		Short: "Broadcast message updateChainParams",
 		Args:  cobra.ExactArgs(2),
 		RunE: func(cmd *cobra.Command, args []string) (err error) {
-
 			argChainParams := args[0]
 
 			clientCtx, err := client.GetClientTxContext(cmd)