--- a/x/observer/client/cli/tx_update_gas_price_increase_flags.go
+++ b/x/observer/client/cli/tx_update_gas_price_increase_flags.go
@@ -18,7 +18,6 @@ func CmdUpdateGasPriceIncreaseFlags() *cobra.Command {
 		Short: "Update the gas price increase flags",
 		Args:  cobra.ExactArgs(5),
 		RunE: func(cmd *cobra.Command, args []string) (err error) {
-
 			clientCtx, err := client.GetClientTxContext(cmd)
 			if err != nil {
 				return err