--- a/x/observer/client/cli/tx_update_observer.go
+++ b/x/observer/client/cli/tx_update_observer.go
@@ -19,7 +19,6 @@ func CmdUpdateObserver() *cobra.Command {
 		Short: "Broadcast message add-observer",
 		Args:  cobra.ExactArgs(3),
 		RunE: func(cmd *cobra.Command, args []string) (err error) {
-
 			clientCtx, err := client.GetClientTxContext(cmd)
 			if err != nil {
 				return err