--- a/x/observer/client/querytests/suite.go
+++ b/x/observer/client/querytests/suite.go
@@ -51,7 +51,6 @@ func (s *CliTestSuite) SetupSuite() {
 	s.network = net
 	_, err = s.network.WaitForHeight(1)
 	s.Require().NoError(err)
-
 }
 
 func (s *CliTestSuite) TearDownSuite() {