--- a/x/observer/keeper/grpc_query_observer.go
+++ b/x/observer/keeper/grpc_query_observer.go
@@ -46,5 +46,4 @@ func (k Keeper) ObserverSet(
 	return &types.QueryObserverSetResponse{
 		Observers: observerSet.ObserverList,
 	}, nil
-
 }