--- a/x/observer/keeper/nonce_to_cctx.go
+++ b/x/observer/keeper/nonce_to_cctx.go
@@ -51,9 +51,7 @@ func (k Keeper) GetAllNonceToCctx(ctx sdk.Context) (list []types.NonceToCctx) {
 		if err == nil {
 			list = append(list, val)
 		}
-
 	}
 
 	return
-
 }