--- a/zetaclient/chains/evm/observer/observer.go
+++ b/zetaclient/chains/evm/observer/observer.go
@@ -481,7 +481,6 @@ func (ob *Observer) WatchGasPrice() {
 }
 
 func (ob *Observer) PostGasPrice() error {
-
 	// GAS PRICE
 	gasPrice, err := ob.evmClient.SuggestGasPrice(context.TODO())
 	if err != nil {