# Description

Previously, `make fmt-import fmt-golines` was really slow. I took a look and improved this because I was a bit upset waiting for minutes each time I want to format the code 😅 . Now it's **x45 times faster** (2.5 minutes vs 3.5 seconds on M3 Pro)

- [x] Remove usage of `goimports-reviser` 
- [x] Update `golangcli-lint` in CI
- [x] Implement efficient `make fmt`
- [x] Add a couple of helpful linters
- [x] Fix/Lint the codebase

Main changes are in `scripts/fmt.sh` and `.golangcli.yml`

### Before
```shell
$ time make fmt-import fmt-golines
# ....
make fmt-import fmt-golines  167.77s user 838.34s system 676% cpu 2:28.62 total
```

### After
```shell
$ time make fmt
# ...
make fmt  3.45s user 0.63s system 297% cpu 1.370 total
```